refactoring the google auth
